\name{ci.yhat}
\alias{ci.yhat}
\title{Compute CI}
\description{
  This function retrieves the proper elements from boot.ci.
}
\usage{
  ci.yhat(bty, CI)
}
\arguments{
  \item{bty}{
    Type of CI
    }
  \item{CI}{
    CI
    }
} % end arguments

\details{
  This function retrieves the proper elements from boot.ci.
}
\value{
This function returns the proper elements from boot.ci.
}
\references{
Nimon, K., & Oswald, F. L. (2013). Understanding the results of multiple linear regression: Beyond standardized regression coefficients. \emph{Organizational Research Methods}, \emph{16}, 
650-674.
}
\author{ Kim Nimon <kim.nimon@gmail.com>}
\note{This function is internal to the yhat package and not intended to be an end-user function.}
